fixed login, signup, logout and users
